requirements.txt
fastapi[standard]==0.116.1
python-jose[cryptography]==3.5.0
motor==3.7.1
python-multipart==0.0.20
pandas==2.3.1
bcrypt==4.3.0
python-dotenv==1.1.1
